move story files to stories subfolder
