+++ ui/packages/shared/components/src/Input/stories/Input.stories.tsx
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Input from './';
+import Input from '..';
 
 export default {
   component: Input,
