+++ ui/packages/shared/components/src/Tab/stories/Tab.stories.tsx
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Tab from './';
+import Tab from '..';
 
 export default {
   component: Tab,
